# apps/desktop/openapi.json
{
  "components": {
    "schemas": {
...
      "CodeSessionOut": {
        "properties": {
          "exchanges": {
            "items": {
              "$ref": "#/components/schemas/CodeExchangeOut"
            },
            "title": "Exchanges",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace",
          "exchanges"
        ],
        "title": "CodeSessionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PostureFacts": {
        "description": "What is true right now, for the UI to render as one sentence.\n\nStructured rather than prose because the sentence has to exist in nine languages, and a server\nthat returned English would quietly make this the one untranslated line on the screen.",
        "properties": {
          "fell_back_to_host": {
            "default": false,
            "title": "Fell Back To Host",
            "type": "boolean"
          },
          "pauses": {
            "enum": [
              "always",
              "tainted",
              "never"
            ],
            "title": "Pauses",
            "type": "string"
          },
          "shell": {
            "enum": [
              "none",
              "isolated",
              "host",
              "asks",
              "refused"
            ],
            "title": "Shell",
            "type": "string"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          },
          "writes": {
            "enum": [
              "nothing",
              "workspace"
            ],
            "title": "Writes",
            "type": "string"
          }
        },
        "required": [
          "writes",
          "workspace",
          "shell",
          "pauses"
        ],
        "title": "PostureFacts",
        "type": "object"
      },
      "PostureQuery": {
        "description": "Ask what a posture would mean, without committing to it \u2014 the selectors' live preview.",
        "properties": {
          "approval": {
            "default": "suspicious",
            "enum": [
              "always",
              "suspicious",
              "never"
            ],
            "title": "Approval",
            "type": "string"
          },
          "reach": {
            "default": "workspace",
            "enum": [
              "read_only",
              "workspace",
              "workspace_shell"
            ],
            "title": "Reach",
            "type": "string"
          },
          "surface": {
            "default": "run",
            "title": "Surface",
            "type": "string"
          },
          "workspace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace"
          }
        },
        "title": "PostureQuery",
        "type": "object"
      },
...
      "RoleModels": {
        "description": "One model slug per role. ``None`` anywhere means \"the run's default model\".\n\n``verify`` is deliberately absent rather than nullable \u2014 offering a field for it would imply a\nchoice exists, and the whole value of an executable verifier is that there isn't one.",
        "properties": {
          "edit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edit"
          },
          "explore": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explore"
          },
          "fuse_plan": {
            "default": false,
            "title": "Fuse Plan",
            "type": "boolean"
          },
          "fuse_review": {
            "default": false,
            "title": "Fuse Review",
            "type": "boolean"
          },
          "plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan"
          },
          "review": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review"
          }
        },
        "title": "RoleModels",
        "type": "object"
      },
      "RolesQuery": {
        "description": "Ask which model each role would run on, without committing to it.",
        "properties": {
          "profile": {
            "default": "balanced",
            "enum": [
              "economy",
              "balanced",
              "max"
            ],
            "title": "Profile",
            "type": "string"
          }
        },
        "title": "RolesQuery",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Chimera Desktop API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/code/posture": {
      "post": {
        "description": "What the chosen posture MEANS on this machine, right now.\n\nA POST rather than a GET because it reports the live state of the sandbox rather than a\nstored resource, and because caching this answer is precisely the bug: a Docker daemon that\ndied since the last call must change the answer, not be served from a cache.",
        "operationId": "code_posture_api_code_posture_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostureQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostureFacts"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Code Posture"
      }
    },
    "/api/code/roles": {
      "post": {
        "description": "The concrete model slugs a profile resolves to, so the UI can show them.\n\nResolved server-side rather than mirrored in the frontend: the tiers already honour the\nuser's cost mode and explicit per-tier settings, and a second copy of that resolution in\nTypeScript would display a model the run does not actually use.",
        "operationId": "code_roles_api_code_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolesQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleModels"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Code Roles"
      }
    },
...
    "/api/code/sessions/{session_id}": {
      "delete": {
        "description": "Forget a conversation. An unknown id is ``{ok: false}`` with a 200, not a 404 \u2014 that is\nexactly the state a second click on Clear hits, and it is not an error.",
        "operationId": "delete_code_session_api_code_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Code Session Api Code Sessions  Session Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Code Session"
      },
      "get": {
        "description": "A stored conversation as the exchanges a person had, ready to render.\n\nResuming used to show an empty screen while the agent silently carried the whole history \u2014\nthe worst of both, because the next question worked for reasons the user could not see.\n\nAn unknown id returns an empty conversation rather than a 404: the store treats a missing\nfile as the ordinary first-turn case, and a screen that errors on a session someone just\ndeleted in another window would be reporting a race as a fault.",
        "operationId": "get_code_session_api_code_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Code Session"
      }
    },
    "/api/code/turn": {
      "post": {
        "operationId": "code_turn_api_code_turn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Code Turn"
      }
    },
...
  }
}